Update search functionality and code refactoring
Updated the search functionality in the omni-wp-api plugin by replacing the previous and next buttons with a load more button. The code has been refactored for better readability and efficiency. Also, several strings have been made translatable for internationalization. Some issues of hardcoding in the loading of assets have been taken care of by defining the plugin version as a constant.