Remove escape function and uncomment cache functions
Removed the esc_attr function which was not necessary in the settings-page.php file. In addition, uncommented the cache check in the ClassPublic.php file that tried to fetch data from cached results before making a new search request, this will help to reduce the server load. The unnecessary newline in the api.php file was also removed.