Updated search functionality to trigger on form submission
The code updates were mainly in JavaScript and PHP files. A key change was to trigger the search function when the search form is submitted, instead of on a button click. The form input type was also changed from "button" to "submit". Other modifications are minor formatting and code order switches.