Update plugin code for performance and usability improvements
In this commit, we have made several updates to enhance the functionality and user experience of the Omni WP API plugin. Modifications were made to the sync_data() function in the ClassAdmin.php and various changes were made in the api.php file to improve performance, including increasing the HTTP request timeout to 20000 milliseconds. Also, the input placeholder in the search-form.php was updated for better user instructions. Lastly, a bulk save functionality was added via a new hook in omni-wp-api.php.