Add project selection and project creation functionality
Several adjustments and new features have been added to the code. We have incorporated the logic to handle a scenario when an existing project is selected. Additionally, we have added a function to retrieve a list of projects, handle project selection, and create a new project. The modal for project creation is also included. Moreover, the condition in the re-indexing meta-data has been modified and sanitization for received data has been included.